+++ sysroot/usr/include/kernel/dev/rtc/rtc.hpp
@@ -17,11 +17,11 @@
 #define RTC_CURRENT_YEAR    2019    // Needs to be updated every year!
 #define RTC_CURRENT_CENTURY 0       // Needs to be updated every century!
 
-extern uint8_t px_rtc_second; // Current UTC second
-extern uint8_t px_rtc_minute; // Current UTC minute
-extern uint8_t px_rtc_hour;   // Current UTC hour
-extern uint8_t px_rtc_day;    // Current UTC day (not reliable)
-extern uint8_t px_rtc_month;  // Current UTC month
+extern uint32_t px_rtc_second; // Current UTC second
+extern uint32_t px_rtc_minute; // Current UTC minute
+extern uint32_t px_rtc_hour;   // Current UTC hour
+extern uint32_t px_rtc_day;    // Current UTC day (not reliable)
+extern uint32_t px_rtc_month;  // Current UTC month
 extern uint32_t px_rtc_year;  // Current UTC year
 
 /**
